refactor: centralize post types
