+++ app/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import Header from "../components/Header";
-import PostCard, { Post } from "../components/PostCard";
+import PostCard from "../components/PostCard";
 import StoryDeckModal from "../components/StoryDeckModal";
 import { posts as samplePosts } from "../data/posts";
 import { useState, useMemo } from "react";
@@ -9,6 +9,7 @@ import CategoryFilter, { Category } from "../components/CategoryFilter";
 import SkeletonCard from "../components/SkeletonCard";
 import { useInfinite } from "../hooks/useInfinite";
 import { motion } from "framer-motion";
+import type { Post } from "../types/post";
 
 export default function Page() {
   const [open, setOpen] = useState(false);
